movied calendar to its own component rather than having all the code in the profile page
