+++ client/src/components/index.js
@@ -3,3 +3,4 @@ export { default as Header } from './Header';
 export { default as Nav } from './Nav';
 export { default as LoginForm } from './LoginForm';
 export { default as SignupForm } from './SignupForm';
+export { default as Calendar } from './Calendar';
