--- client/src/pages/Profile/style.css
@@ -1,21 +1,0 @@
-table {
-    border-collapse: collapse;
-    width: 30%;
-  }
-  
-  th,
-  td {
-    border: 1px solid black;
-    padding: 8px;
-    text-align: center;
-  }
-  
-  #calendarHeader {
-    display: flex;
-  }
-  
-  #currentMonth,
-  #calendarHeader button {
-    display: inline-block;
-    margin: 5px
-  }
